# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HeyCreator API",
    "description": "Influencer discovery and campaign management platform API. Connects brands with creators across Instagram, TikTok, YouTube, Twitter, and Facebook.",
    "version": "1.0.0",
    "contact": {
      "name": "HeyCreator Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/api/profiles/{id}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Get profile details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Profile with full details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/UnifiedProfile" }
                  }
                }
              }
            }
          },
          "404": { "description": "Profile not found" }
        }
      },
      "put": {
        "tags": ["Profiles"],
        "summary": "Update profile",
        "description": "Update a unified profile. Supports adding/removing linked accounts and updating profile details.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "addAccounts": { "type": "array", "items": { "type": "object" } },
                  "removeAccounts": { "type": "array", "items": { "type": "string" } },
                  "updateDetails": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Profile updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "Profile not found" }
        }
      },
      "delete": {
        "tags": ["Profiles"],
        "summary": "Soft-delete profile",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Profile deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } },
          "404": { "description": "Profile not found" }
        }
      }
    },
    "/api/influencer/profiles": {
      "get": {
        "tags": ["Profiles"],
        "summary": "List influencer profiles",
        "description": "Fetches all influencer profiles for the authenticated user. Supports filtering by status.",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["active", "archived", "deleted"] } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Influencer profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/UnifiedProfile" } },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID Token",
        "description": "Firebase Authentication ID token. Obtain by calling `firebase.auth().currentUser.getIdToken()`."
      },
      "UserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Development fallback: Firebase UID passed directly as header. Not for production use."
      }
    },
    "schemas": {
      "Platform": {
        "type": "string",
        "enum": ["instagram", "tiktok", "youtube", "twitter", "facebook"]
      },
...
      "UnifiedProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Firestore document ID" },
          "userId": { "type": "string" },
          "primaryAccountKey": { "type": "string", "description": "platform:username dedup key" },
          "displayName": { "type": "string" },
          "bio": { "type": "string" },
          "location": { "type": "string" },
          "categories": { "type": "array", "items": { "type": "string" } },
          "avatarUrl": { "type": "string" },
          "linkedAccounts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "platform": { "$ref": "#/components/schemas/Platform" },
                "username": { "type": "string" },
                "followerCount": { "type": "integer" },
                "verified": { "type": "boolean" },
                "profileUrl": { "type": "string" }
              }
            }
          },
          "combinedMetrics": {
            "type": "object",
            "properties": {
              "totalFollowers": { "type": "integer" },
              "averageEngagementRate": { "type": "number" },
              "totalReach": { "type": "integer" }
            }
          },
          "influenceScore": { "type": "number" },
          "estimatedPrice": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "isActive": { "type": "boolean" }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "type": "object" },
          "count": { "type": "integer" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "details": { "type": "string" }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized — missing or invalid authentication token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
    }
  }
}